Create currency task

+++ db/migrate/20211210134959_create_currencies.rb
@@ -2,7 +2,7 @@ class CreateCurrencies < ActiveRecord::Migration[6.1]
   def change
     create_table :currencies do |t|
       t.string :iso_code, null: false
-      t.float :rate, null: false
+      t.decimal :rate, null: false
       t.boolean :is_base, default: false
 
       t.timestamps
